# AuthService/api-doc.json
{"openapi":"3.0.1","info":{"title":"test","description":"testing","version":"0.0.0"},"servers":[{"url":"http://localhost:9004","description":"Generated server url"}],"paths":{"/profile":{"get":{"tags":["profile-controller"],"operationId":"listAllFormsOfMetadata_1","responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/RepresentationModelObject"}}}}}}},"/profile/roles":{"get":{"tags":["profile-controller"],"operationId":"descriptor_1_1_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}},"application/alps+json":{"schema":{"type":"string"}},"application/schema+json":{"schema":{"$ref":"#/components/schemas/JsonSchema"}}}}}}},"/profile/statuses":{"get":{"tags":["profile-controller"],"operationId":"descriptor_1_1_2","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}},"application/alps+json":{"schema":{"type":"string"}},"application/schema+json":{"schema":{"$ref":"#/components/schemas/JsonSchema"}}}}}}},"/profile/users":{"get":{"tags":["profile-controller"],"operationId":"descriptor_1_1_3","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}},"application/alps+json":{"schema":{"type":"string"}},"application/schema+json":{"schema":{"$ref":"#/components/schemas/JsonSchema"}}}}}}},"/roles":{"get":{"tags":["role-entity-controller"],"description":"get-role","operationId":"getCollectionResource-role-get_1","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"minimum":0,"type":"integer","default":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"minimum":1,"type":"integer","default":20}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelRole"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelRole"}},"text/uri-list":{"schema":{"type":"string"}}}}}},"post":{"tags":["role-entity-controller"],"description":"create-role","operationId":"postCollectionResource-role-post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleRequestBody"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelRole"}}}}}}},"/roles/{id}":{"get":{"tags":["role-entity-controller"],"description":"get-role","operationId":"getItemResource-role-get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelRole"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["role-entity-controller"],"description":"update-role","operationId":"putItemResource-role-put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelRole"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelRole"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["role-entity-controller"],"description":"delete-role","operationId":"deleteItemResource-role-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["role-entity-controller"],"description":"patch-role","operationId":"patchItemResource-role-patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoleRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelRole"}}}},"204":{"description":"No Content"}}}},"/statuses":{"get":{"tags":["status-entity-controller"],"description":"get-status","operationId":"getCollectionResource-status-get_1","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"minimum":0,"type":"integer","default":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"minimum":1,"type":"integer","default":20}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelStatus"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelStatus"}},"text/uri-list":{"schema":{"type":"string"}}}}}},"post":{"tags":["status-entity-controller"],"description":"create-status","operationId":"postCollectionResource-status-post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusRequestBody"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelStatus"}}}}}}},"/statuses/{id}":{"get":{"tags":["status-entity-controller"],"description":"get-status","operationId":"getItemResource-status-get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelStatus"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["status-entity-controller"],"description":"update-status","operationId":"putItemResource-status-put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelStatus"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelStatus"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["status-entity-controller"],"description":"delete-status","operationId":"deleteItemResource-status-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["status-entity-controller"],"description":"patch-status","operationId":"patchItemResource-status-patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StatusRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelStatus"}}}},"204":{"description":"No Content"}}}},"/users":{"get":{"tags":["user-entity-controller"],"description":"get-user","operationId":"getCollectionResource-user-get_1","parameters":[{"name":"page","in":"query","description":"Zero-based page index (0..N)","required":false,"schema":{"minimum":0,"type":"integer","default":0}},{"name":"size","in":"query","description":"The size of the page to be returned","required":false,"schema":{"minimum":1,"type":"integer","default":20}},{"name":"sort","in":"query","description":"Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.","required":false,"schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelUser"}},"application/x-spring-data-compact+json":{"schema":{"$ref":"#/components/schemas/PagedModelEntityModelUser"}},"text/uri-list":{"schema":{"type":"string"}}}}}},"post":{"tags":["user-entity-controller"],"description":"create-user","operationId":"postCollectionResource-user-post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestBody"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelUser"}}}}}}},"/users/search/findByLogin":{"get":{"tags":["user-search-controller"],"operationId":"executeSearch-user-get","parameters":[{"name":"login","in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelUser"}}}},"404":{"description":"Not Found"}}}},"/users/{id}":{"get":{"tags":["user-entity-controller"],"description":"get-user","operationId":"getItemResource-user-get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelUser"}}}},"404":{"description":"Not Found"}}},"put":{"tags":["user-entity-controller"],"description":"update-user","operationId":"putItemResource-user-put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelUser"}}}},"201":{"description":"Created","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelUser"}}}},"204":{"description":"No Content"}}},"delete":{"tags":["user-entity-controller"],"description":"delete-user","operationId":"deleteItemResource-user-delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"204":{"description":"No Content"},"404":{"description":"Not Found"}}},"patch":{"tags":["user-entity-controller"],"description":"patch-user","operationId":"patchItemResource-user-patch","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestBody"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/EntityModelUser"}}}},"204":{"description":"No Content"}}}},"/auth/register":{"post":{"tags":["auth-controller"],"description":"Registration Endpoint","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegistrationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/auth/login":{"post":{"tags":["auth-controller"],"description":"Login Endpoint","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/auth/activate":{"post":{"tags":["auth-controller"],"description":"Activation Endpoint","operationId":"activate","parameters":[{"name":"login","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/hal+json":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}}},"components":{"schemas":{"AbstractJsonSchemaPropertyObject":{"type":"object","properties":{"title":{"type":"string"},"readOnly":{"type":"boolean"}}},"Item":{"type":"object","properties":{"type":{"type":"string"},"properties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/AbstractJsonSchemaPropertyObject"}},"requiredProperties":{"type":"array","items":{"type":"string"}}}},"JsonSchema":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"properties":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/AbstractJsonSchemaPropertyObject"}},"requiredProperties":{"type":"array","items":{"type":"string"}},"definitions":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Item"}},"type":{"type":"string"},"$schema":{"type":"string"}}},"Links":{"type":"object","additionalProperties":{"$ref":"#/components/schemas/Link"}},"RepresentationModelObject":{"type":"object","properties":{"_links":{"$ref":"#/components/schemas/Links"}}},"EntityModelStatus":{"type":"object","properties":{"name":{"type":"string"},"_links":{"$ref":"#/components/schemas/Links"}}},"PageMetadata":{"type":"object","properties":{"size":{"type":"integer","format":"int64"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int64"}}},"PagedModelEntityModelStatus":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"statuses":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelStatus"}}}},"_links":{"$ref":"#/components/schemas/Links"},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"EntityModelRole":{"type":"object","properties":{"name":{"type":"string"},"_links":{"$ref":"#/components/schemas/Links"}}},"PagedModelEntityModelRole":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"roles":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelRole"}}}},"_links":{"$ref":"#/components/schemas/Links"},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"EntityModelUser":{"type":"object","properties":{"email":{"type":"string"},"login":{"type":"string"},"password":{"type":"string"},"role":{"type":"integer","format":"int64"},"statusId":{"type":"integer","format":"int64"},"_links":{"$ref":"#/components/schemas/Links"}}},"PagedModelEntityModelUser":{"type":"object","properties":{"_embedded":{"type":"object","properties":{"users":{"type":"array","items":{"$ref":"#/components/schemas/EntityModelUser"}}}},"_links":{"$ref":"#/components/schemas/Links"},"page":{"$ref":"#/components/schemas/PageMetadata"}}},"RoleRequestBody":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"StatusRequestBody":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"UserRequestBody":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"email":{"type":"string"},"login":{"type":"string"},"password":{"type":"string"},"role":{"type":"integer","format":"int64"},"statusId":{"type":"integer","format":"int64"}}},"RegistrationRequest":{"type":"object","properties":{"email":{"type":"string"},"login":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"}}},"AuthResponse":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}},"LoginRequest":{"type":"object","properties":{"login":{"type":"string"},"password":{"type":"string"}}},"ActivationRequest":{"type":"object","properties":{"firstname":{"type":"string"},"lastname":{"type":"string"}}},"Link":{"type":"object","properties":{"href":{"type":"string"},"hreflang":{"type":"string"},"title":{"type":"string"},"type":{"type":"string"},"deprecation":{"type":"string"},"profile":{"type":"string"},"name":{"type":"string"},"templated":{"type":"boolean"}}}}}}