# AuthService/src/main/java/com/example/AuthService/Entities/ActivationRequest.java
package com.example.AuthService.Entities;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class ActivationRequest {

    private String firstname;
    private String lastname;


}