# UserService/src/main/java/com/example/UserService/Models/AuthResponse.java
//package com.example.UserService.Models;
//
//import lombok.AllArgsConstructor;
//import lombok.Builder;
//import lombok.Data;
//import lombok.NoArgsConstructor;
//
//@Data
//@AllArgsConstructor
//@NoArgsConstructor
//@Builder
//public class AuthResponse {
//    private String accessToken;
//    private String refreshToken;
//}